correct plotting issue in sub2AvgBrain
When plotting stereo-EEG cases who have electrodes in both hemispheres, the plots that are generated by the function misplace the electrodes on fsaverage (see attached PPT). Despite this plotting error, I think that the fsaverage electrode coordinates themselves are correct, since plotting the electrodes again on fsaverage using the output from sub2AvgBrain performs as expected.

I have slightly modified the code in the plotting section of sub2AvgBrain to correct the issue.